rename: RegisterByCategory -> InformationByCategory 로 변경
